<==== Mercurial data ====>
Branch: dev_3.0.0_tcpReverseConnect

+++ common/src/network/h264_rtp_parser.cpp
@@ -556,7 +556,8 @@ bool CLH264RtpParser::processData(
                 --curPtr;
                 *curPtr = nalUnitType;
             }
-            m_chunks.emplace_back(curPtr-rtpBufferBase, bufferEnd - curPtr, m_packetPerNal == 0);
+            m_chunks.emplace_back(
+                (int) (curPtr-rtpBufferBase), bufferEnd - curPtr, m_packetPerNal == 0);
             m_videoFrameSize += bufferEnd - curPtr + (m_packetPerNal == 0 ? 4 : 0);
             break;
         }
